+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the Civo CLI will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2019-09-21
+### Changed
+- `civo kubernetes config` will show an error (and exit with a non-zero status) if the config isn't available from the server
+
+### Added
+- Add `civo update` command to automatically update your `civo_cli` Ruby gem to the latest version.
+
 ## [0.4.0] - 2019-09-17
 ### Changed
 - Now the capability exists for pre-installed applications in a cluster, ensure they aren't manually installed